refactor is_supported
